pyproject.toml
[tool.poetry]
name = "cw-drf"
version = "0.1.0"
description = ""
authors = ["m.romantsov <romancov_dzr@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
isort = "^5.13.2"
django-cors-headers = "^4.6.0"
celery = "^5.4.0"
redis = "^5.2.0"
python-telegram-bot = "^21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
